+++ NEWS.md
@@ -1,6 +1,10 @@
 # DataFrames.jl v1.5 Release Notes
 
-* add `allunique` and allow transformations in `cols` argument of `describe`
+## New functionalities
+
+* Add `Iterators.partition` support
+   ([#3212](https://github.com/JuliaData/DataFrames.jl/pull/3212))
+* Add `allunique` and allow transformations in `cols` argument of `describe`
   and `nonunique` when working with `SubDataFrame`
   ([3232](https://github.com/JuliaData/DataFrames.jl/pull/3232))